Fixed all spellings of "Vampyre" to "Vampire". We now *USE* the max class count fron the race table. Humans will now be able to level 4 mortal classes, all other races 3 (for now). Gain/Worth were updated to only display gainable classes with this. Rlist was updated to display max class count. There are no restrictions based on class order, so players will still enter 5 classes at creation and have the option to mix/match as they see fit down the road; if they are willing to take the skill penalties for using their 4th or 5th ranked classes. Closes Mantis 16.

+++ src/h/ack.h
@@ -383,7 +383,7 @@ struct race_type
     short race_con;
     int race_flags;   /* flags for the various racial stuff    */
     bool wear_locs[MAX_WEAR];  /* on/off for each wear loc */
-    int classes;   /* Number of classes for race   */
+    short classes;   /* Number of classes for race   */
     short limit[MAX_CLASS];   /* Max for each class */
     char *comment; /* comments shown for new players */
     char *skill1;
